// frontend/src/api/generated/model/pointSourceSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * astroimage
 * OpenAPI spec version: 0.1.0
 */

export interface PointSourceSchema {
  source_id: number;
  rank: number;
  xcentroid: number;
  ycentroid: number;
  snr: number;
  relevance_score: number;
  peak?: number | null;
  flux?: number | null;
  object_type?: 'point';
}